Refactor template file structure
