+++ src/templates/CentralTemplateProvider.ts
