+++ src/templates/dotnet/executeDotnetTemplateCommand.ts
