+++ src/templates/script/parseScriptTemplates.ts
